mnogunik/mnogunik/requirements.txt
pandas
openpyxl
pillow
requests
chardet
python-dateutil